% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montgomeryDecomposition.R
\name{montgomeryDecompostion}
\alias{montgomeryDecompostion}
\title{Title}
\usage{
montgomeryDecompostion(.l, .decFormula, .output = "default")
}
\arguments{
\item{.l}{A list with the variables that create the variable}

\item{.decFormula}{A string that contains}

\item{.output}{The desired display form of the decomposition results}
}
\value{

}
\description{
Title
}
\examples{
lofVars <- list("A" = list("t0" = matrix(runif(8), ncol=4), "t1" = matrix(runif(8), ncol=4)),
                "C" = list("t0" = matrix(runif(27), nrow=3), "t1" = matrix(runif(27), nrow=3)),
                "B" = list("t0" = matrix(runif(12), nrow=4), "t1" = matrix(runif(12), nrow=4)),
                "D" = list("t0" = matrix(runif(27), nrow =3), "t1"= matrix(runif(27), nrow=3)))

lofVars \%>\% montgomeryDecomposition("A\%*\%B\%*\%C", "all")

lofVars \%>\% montgomeryDecomposition("A\%*\%B\%*\%C*D", "all")


}
